+++ sklearn/base.py
@@ -14,32 +14,14 @@ import numpy as np
 
 from . import __version__
 from ._config import get_config
-from .utils import _IS_32BIT
+from .utils import (
+    _DEFAULT_TAGS,
+    _IS_32BIT,
+)
 from .utils.validation import check_X_y
 from .utils.validation import check_array
 from .utils._estimator_html_repr import estimator_html_repr
 from .utils.validation import _deprecate_positional_args
-
-_DEFAULT_TAGS = {
-    'non_deterministic': False,
-    'requires_positive_X': False,
-    'requires_positive_y': False,
-    'X_types': ['2darray'],
-    'poor_score': False,
-    'no_validation': False,
-    'multioutput': False,
-    "allow_nan": False,
-    'stateless': False,
-    'multilabel': False,
-    '_skip_test': False,
-    '_xfail_checks': False,
-    'multioutput_only': False,
-    'binary_only': False,
-    'requires_fit': True,
-    'preserves_dtype': [np.float64],
-    'requires_y': False,
-    'pairwise': False,
-    }
 
 
 @_deprecate_positional_args
